Refactor typedefs into separate file

+++ src/framebuffer/framebuffer.cpp
@@ -20,7 +20,7 @@ namespace Render
    * @param width The new width of the window.
    * @param height The new height of the window.
    */
-  void ResizeFramebuffer(Framebuffer* buffer, int width, int height)
+  void ResizeFramebuffer(Framebuffer* buffer, int16 width, int16 height)
   {
     if (buffer->bitmap)
     {
@@ -30,24 +30,24 @@ namespace Render
     buffer->width = width;
     buffer->height = height;
 
-    const int bytesPerPixel = 4;
-    int bitmapSize = (width * height) * bytesPerPixel;
+    const int8 bytesPerPixel = 4;
+    int32 bitmapSize = (width * height) * bytesPerPixel;
     
     buffer->bitmap = VirtualAlloc(0, bitmapSize, MEM_COMMIT, PAGE_READWRITE);
   }
   
   // NOTE: This will change or be removed. No docstring needed.
-  void RenderGradient(Render::Framebuffer* buffer, int xOffset, int yOffset)
+  void RenderGradient(Render::Framebuffer* buffer, int32 xOffset, int32 yOffset)
   {
     int pitch = buffer->width*4;
-    uint8_t* row = (uint8_t*)buffer->bitmap;
+    uint8* row = (uint8*)buffer->bitmap;
     for(int y = 0; y < buffer->height; ++y)
     {
-      uint32_t* pixel = (uint32_t*)row;
+      uint32* pixel = (uint32*)row;
       for(int x = 0; x < buffer->width; ++x)
       {
-        *pixel = (uint8_t)(x+xOffset) << 16 | 
-                (uint8_t)(y+yOffset) << 8 | 
+        *pixel = (uint8)(x+xOffset) << 16 | 
+                (uint8)(y+yOffset) << 8 | 
                 255;
         ++pixel;
       }
